+++ CHANGELOG.md
@@ -85,6 +85,9 @@ Adding a new version? You'll need three changes:
   [#4749](https://github.com/Kong/kubernetes-ingress-controller/pull/4749)
 - Removed Knative support.
   [#4748](https://github.com/Kong/kubernetes-ingress-controller/pull/4748)
+- Removed support for deprecated `KongIngress` fields: `Proxy` and `Route`. Respective
+  `Service` or `Ingress` annotations should be used instead. See [KIC Annotations reference].
+  [#4760](https://github.com/Kong/kubernetes-ingress-controller/pull/4760)
 
 ### Fixed
 
@@ -98,6 +101,8 @@ Adding a new version? You'll need three changes:
 - Get rid of deprecation warning in logs for unsupported label `global: true` for `KongPlugin`,
   it'll be treated as any other label without a special meaning.
   [#4737](https://github.com/Kong/kubernetes-ingress-controller/pull/4737)
+
+[KIC Annotations reference]: https://docs.konghq.com/kubernetes-ingress-controller/latest/references/annotations/
 
 ## 2.12.0